Improve button component & implement Search feature

+++ springboot-auto-management/src/main/java/com/auto/management/model/Car.java
@@ -7,7 +7,7 @@ import javax.persistence.*;
 public class Car {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Column(name = "year")
     private int year;
@@ -29,28 +29,12 @@ public class Car {
         this.model = model;
     }
 
-    public int getId() {
-        return id;
-    }
-    public void setId(int id) {
-        this.id = id;
-    }
-    public int getYear() {
-        return year;
-    }
-    public void setYear(int year) {
-        this.year = year;
-    }
-    public String getMake() {
-        return make;
-    }
-    public void setMake(String make) {
-        this.make = make;
-    }
-    public String getModel() {
-        return model;
-    }
-    public void setModel(String model) {
-        this.model = model;
-    }
+    public Long getId() { return id; }
+    public void setId(Long id) { this.id = id; }
+    public int getYear() { return year; }
+    public void setYear(int year) { this.year = year; }
+    public String getMake() { return make; }
+    public void setMake(String make) { this.make = make; }
+    public String getModel() { return model; }
+    public void setModel(String model) { this.model = model; }
 }
